requirements.txt
attrs==17.2.0
beautifulsoup4==4.6.0